refactor: new abstract ms class
